+++ version_3.0/.gitignore
@@ -1,0 +1,1 @@
+***.mp4
