+++ version_3.0/experimentes/stream_static_video_from_server/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Streaming Video</title>
+  </head>
+  <body>
+    <video width="640" height="360" controls>
+      <source src="http://localhost:731/video" type="video/mp4" />
+      Your browser does not support the video tag.
+    </video>
+  </body>
+</html>
