+++ version_3.0/server.js
@@ -1,5 +1,13 @@
 const WebSocket = require("ws");
 const http = require("http");
+const express = require("express");
+const fs = require("fs");
+
+const app = express();
+const port = 731;
+
+const videoPath = "movie.mp4";
+const videoSize = fs.statSync(videoPath).size;
 
 // Create an HTTP server
 const server = http.createServer((req, res) => {
@@ -36,4 +44,34 @@ wss.on("connection", (ws) => {
 const PORT = process.env.PORT || 3000;
 server.listen(PORT, () => {
   console.log(`Server is listening on port ${PORT}`);
+});
+
+app.get("/video", (req, res) => {
+  const range = req.headers.range;
+
+  if (!range) {
+    res.status(400).send("Range header is required");
+    return;
+  }
+
+  const parts = range.replace(/bytes=/, "").split("-");
+  const start = parseInt(parts[0], 10);
+  const end = parts[1] ? parseInt(parts[1], 10) : videoSize - 1;
+
+  const chunksize = end - start + 1;
+
+  const file = fs.createReadStream(videoPath, { start, end });
+  const head = {
+    "Content-Range": `bytes ${start}-${end}/${videoSize}`,
+    "Accept-Ranges": "bytes",
+    "Content-Length": chunksize,
+    "Content-Type": "video/mp4",
+  };
+
+  res.writeHead(206, head);
+  file.pipe(res);
+});
+
+app.listen(port, () => {
+  console.log(`Server is running on http://localhost:${port}`);
 });
