--- version_3.0/stream_static_video_from_server/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Streaming Video</title>
-  </head>
-  <body>
-    <video width="640" height="360" controls>
-      <source src="http://13.53.212.61:3000/video" type="video/mp4" />
-      Your browser does not support the video tag.
-    </video>
-  </body>
-</html>
